refactor: consolidate monorepo tests + add new Jest tests to make sure public packages don't use peerDependencies or devDependencies
