+++ jest.config.js
@@ -31,7 +31,10 @@ module.exports = {
   ],
   globalSetup: './jest-global-setup.js',
   globalTeardown: './jest-global-teardown.js',
-  setupFilesAfterEnv: ['./jest-setup-files-after-env.js'],
+  setupFilesAfterEnv: [
+    './jest-setup-files-after-env.js',
+    'jest-expect-message',
+  ],
   snapshotSerializers: ['jest-serializer-html'],
   reporters: ['default', '<rootDir>/scripts/report-jest-screenshots.js'],
   // Notify not working correctly; we want to only get a notification when tests fail, and then get ONE success notificaiton after it passes
