+++ package.json
@@ -42,7 +42,7 @@
     "test": "npm-run-all --parallel test:js test:php test:monorepo",
     "test:js": "NODE_ENV='test' jest --all --colors",
     "test:js:update": "yarn run test:js -u",
-    "test:monorepo": "node scripts/monorepo-tests.js",
+    "test:monorepo": "npx jest __tests__/monorepo.test.js -c jest.config.quick.js",
     "test:php": "cd packages/core-php -- composer run test",
     "test:e2e:quick-live": " NOW_URL=https://boltdesignsystem.com ./node_modules/.bin/nightwatch --config nightwatch.js --env local",
     "test:e2e:full-live": " NOW_URL=https://boltdesignsystem.com ./node_modules/.bin/nightwatch --config nightwatch.js --env chrome,firefox,ie11",
@@ -51,7 +51,6 @@
     "test:e2e:quick-local": "NOW_URL=http://localhost:3000 ./node_modules/.bin/nightwatch --config nightwatch.local.js",
     "test:e2e:full-local": "NOW_URL=http://localhost:3000 ./node_modules/.bin/nightwatch --config nightwatch.local.js --env chrome,firefox,safari",
     "serve": "node packages/servers/default-server",
-    "postbootstrap": "node scripts/monorepo-tests.js",
     "cz": "git-cz",
     "release:rc": "./scripts/release/release-rc.sh",
     "release:beta": "./scripts/release/release-beta.sh",
@@ -59,7 +58,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-push": "yarn run lint",
+      "pre-push": "yarn run lint && npm run test:monorepo",
       "commitmsg": "commitlint -E $GIT_PARAMS"
     }
   },
@@ -77,6 +76,7 @@
     "access": "public"
   },
   "devDependencies": {
+    "chalk": "^2.4.2",
     "document-register-element": "^1.13.1",
     "postinstall-postinstall": "^2.0.0",
     "patch-package": "^6.0.5",
@@ -105,6 +105,7 @@
     "ci-utils": "^0.5.0",
     "cli-highlight": "^2.0.0",
     "express": "^4.16.4",
+    "jest-expect-message": "^1.0.2",
     "jest-dev-server": "^4.0.0",
     "jest-dom": "^3.1.3",
     "jest-environment-node": "^24.5.0",
